Code splitting and clean up
