+++ src/css/home.css
@@ -1,0 +1,24 @@
+.top-bar {
+  display: flex;
+  flex-grow: 0;
+  justify-content: space-between;
+  align-items: center;
+  background-color: var(--primary-color);
+  padding: 0 20px;
+}
+
+.title-container {
+  display: flex;
+  align-items: center;
+  color: #fff;
+  height: 75px;
+}
+
+.title {
+  text-shadow: 2px 2px 4px #000000;
+}
+
+.logo {
+  display: block;
+  max-width: 90px;
+}
