+++ src/helpers/diagramHandler.js
