+++ src/index.js
@@ -1,4 +1,6 @@
 import './style.css';
+import './css/home.css';
+import './css/canvas.css';
 import './assets/idea-logo-small.png';
 import router from './router';
